requirements.txt
django-model-utils>=2.0
django-autoslug==1.9.8
django-ndarray==0.0.3
# Additional requirements go here
